todo/requirements.txt
aniso8601==7.0.0
anytree==2.8.0
asgiref==3.5.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.2
django-cors-headers==3.11.0
django-factory-boy==1.0.0
django-filter==21.1
django-seed==0.3.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.2.2
djoser==2.1.0
drf-yasg==1.20.0
fabric==2.6.0
factory-boy==3.2.1
Faker==12.0.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.3
importlib-metadata==4.10.1
inflection==0.5.1
invoke==1.6.0
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.1.0
mixer==7.2.1
oauthlib==3.2.0
packaging==21.3
paramiko==2.9.2
pathlib2==2.3.6
promise==2.3
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==3.0.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
stringcase==1.2.0
text-unidecode==1.3
toposort==1.7
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
wrapt==1.14.0
zipp==3.7.0
